dynamic jurisdictions display
