+++ arlo-client/src/components/Audit/Setup/Contests/index.test.tsx
@@ -4,7 +4,8 @@ import { toast } from 'react-toastify'
 import { useParams } from 'react-router-dom'
 import { regexpEscape, asyncActRender } from '../../../testUtilities'
 import * as utilities from '../../../utilities'
-import Contests, { IJurisdictions } from './index'
+import Contests from './index'
+import { IJurisdictions } from '../useParticipantsApi'
 import relativeStages from '../_mocks'
 import { contestsInputMocks, contestMocks } from './_mocks'
 import { numberifyContest, IContestNumbered } from '../useContestsApi'
